763. Partition Labels.
