--- src/main/java/com/coding/challenge/greedy/PartitionLabels.java
@@ -1,46 +1,0 @@
-package com.coding.challenge.greedy;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-class PartitionLabels {
-	private static final char FIRST_LETTER = 'a';
-
-	PartitionLabels() {
-		throw new AssertionError();
-	}
-
-	public static void main(String[] args) {
-		final String s1 = "ababcbacadefegdehijhklij";
-		assert Arrays.asList(9, 7, 8).equals(partitionLabels(s1));
-
-		final String s2 = "eccbbbbdec";
-		assert Arrays.asList(10).equals(partitionLabels(s2));
-	}
-
-	static List<Integer> partitionLabels(String s) {
-		final int n = s.length();
-		final int[] charFreq = new int[26];
-		for (char ch : s.toCharArray())
-			charFreq[ch - FIRST_LETTER] = charFreq[ch - FIRST_LETTER] + 1;
-
-		final List<Integer> ans = new ArrayList<>();
-		final Set<Character> part = new HashSet<>();
-		for (int i = 0, j = 0, letters = 0; j < n; j++) {
-			part.add(s.charAt(j));
-			charFreq[s.charAt(j) - FIRST_LETTER] = charFreq[s.charAt(j) - FIRST_LETTER] - 1;
-			if (charFreq[s.charAt(j) - FIRST_LETTER] == 0)
-				letters = letters + 1;
-
-			if (part.size() == letters) {
-				ans.add(j - i + 1);
-				i = j + 1;
-			}
-		}
-
-		return ans;
-	}
-}
